orchestrator/requirements.txt
# WAVE Orchestrator Dependencies
# Phase A: Orchestrator Skeleton

# Core LangGraph
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# State & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database & Cache
sqlalchemy>=2.0.0
redis>=5.0.0
asyncpg>=0.30.0

# Git Operations (Phase B)
pygit2>=1.14.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.28.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
